+++ MillimanAccessPortal/MillimanAccessPortal/src/ts/react/login/login-form.tsx
@@ -171,7 +171,8 @@ export class LoginForm extends Form<{}, LoginFormState> {
 
     // hold for
     this.setState({ awaitingConfirmation: true }, () => {
-      postData('/Account/IsLocalAccount', { username: this.state.data.username })
+      const { username } = this.state.data;
+      postData('/Account/IsLocalAccount', { username })
         .then((response) => {
           if (response.localAccount) {
             this.setState({
@@ -181,7 +182,9 @@ export class LoginForm extends Form<{}, LoginFormState> {
               this.focusPasswordInput();
             });
           } else {
-            window.location.replace(`/Account/RemoteAuthenticate?username=${this.state.data.username}`);
+            const urlParemeters = new URLSearchParams(window.location.search);
+            const returnTo = escape(urlParemeters.has('ReturnUrl') ? urlParemeters.get('ReturnUrl') : '/');
+            window.location.replace(`/Account/RemoteAuthenticate?username=${username}&returnURL=${returnTo}`);
           }
         })
         .catch(() => {