+++ Documentation/ReleaseNotes.md
@@ -10,6 +10,7 @@
 - Added ability to show the Bookmarks pane in PowerBI reports
 - Allow `<iframe>` to render PDFs in a new window
 - Clarified several user messages in the password reset workflow
+- Fixed a bug that prevented SSO users from logging into any view other than Authorized Content
 - Fixed a bug that allowed the password reset form to be displayed using an expired email link
 - Fixed an issue that allowed whitespace on username inputs
 - Fixed the styling on republishing icons